+++ login/src/main/java/login/api/Security/jwt/AuthEntryPointJwt.java
@@ -1,4 +1,4 @@
-package login.lib.security.jwt;
+package login.api.Security.jwt;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
