+++ login/src/main/java/login/api/Security/services/UserDetailsImpl.java
@@ -1,4 +1,4 @@
-package login.lib.security.services;
+package login.api.Security.services;
 
 import login.api.Models.User;
 import com.fasterxml.jackson.annotation.JsonIgnore;
