+++ login/src/main/java/login/api/Security/services/WebSecurityConfig.java
@@ -1,7 +1,5 @@
-package login.lib.security.services;
+package login.api.Security.services;
 
-import login.lib.security.jwt.AuthEntryPointJwt;
-import login.lib.security.jwt.AuthTokenFilter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -15,6 +13,9 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import login.api.Security.jwt.AuthEntryPointJwt;
+import login.api.Security.jwt.AuthTokenFilter;
 
 @Configuration
 @EnableWebSecurity
